pyproject.toml
[project]
name = "simcache"
version = "0.1.0"
description = "Content-addressed simulation cache with reproducible plotting."
requires-python = ">=3.9"
dependencies = [
  "numpy",
  "PyYAML",
]

[project.optional-dependencies]
hdf5 = ["h5py"]
zarr = ["zarr"]

[project.scripts]
simcache = "simcache.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
